% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIO.R
\name{fabio}
\alias{fabio}
\title{The main function to perform TWAS fine-mapping on binary outcomes}
\usage{
fabio(
  X,
  y,
  beta_a = 0,
  beta_b = 0,
  w_step = 6000,
  s_step = 20000,
  save_dir = "."
)
}
\arguments{
\item{X}{Input predicted GReX matrix}

\item{y}{A vector of binary phenotypes of TWAS}

\item{beta_a}{alpha of the prior beta distribution on pi, both a and b are 0 by default, leading to a uniform prior on log pi}

\item{beta_b}{beta of the prior beta distribution on pi, both a and b are 0 by default, leading to a uniform prior on log pi}

\item{w_step}{The number of warm-up steps in MCMC, default = 6000}

\item{s_step}{The number of sampling steps in MCMC, default = 20000}

\item{save_dir}{The directory where the output file will be saved, set to be the working directory by default}
}
\value{
The results will be saved as a csv file
}
\description{
The main function to perform TWAS fine-mapping on binary outcomes
}
\examples{
# example files can be downloaded from: 
# https://www.dropbox.com/scl/fo/fxynm8uvedgvy7ni6hcbt/AAfTQVo89s78DsRNwpBH3lU?rlkey=nbqwrdi2r5y1bbojzf7z8ev7h&st=yz28n4nj&dl=0

library(FABIO)

grex <- data.table::fread('./example_grex.txt.gz')
pheno <- scan('./example_pheno.txt',numeric())
fabio(X=grex, y=pheno, w_step=100, s_step=1000)
# The results will be saved as a file named "FABIO_out.csv"
}
